+++ src/web/mod.rs
@@ -1,1 +1,2 @@
+pub mod custom_parser;
 pub mod find_one;
